refactor(client): extract logic from defect modal
